Fix failing tests
